WIP: Refactor Auth code
