--- VacCheck Web/vaccheck/src/fire.js
@@ -1,33 +1,0 @@
-// // import firebase from "firebase";
-
-// // var firebaseConfig = {
-// //   apiKey: "AIzaSyBp1KweTccbGHPLHDXQMkJhikD0j4dZA_4",
-// //   authDomain: "vaccheck-6a24b.firebaseapp.com",
-// //   projectId: "vaccheck-6a24b",
-// //   storageBucket: "vaccheck-6a24b.appspot.com",
-// //   messagingSenderId: "685620451959",
-// //   appId: "1:685620451959:web:8b1946844a5c2dc1550472",
-// //   measurementId: "G-BG18CF3S0E",
-// // };
-
-// // const fire = initializeApp(firebaseConfig);
-
-// // export default fire;
-
-// import firebase from "firebase";
-// // import { getFirestore } from "firebase/firestore";
-
-// const firebaseConfig = {
-//   apiKey: "AIzaSyBp1KweTccbGHPLHDXQMkJhikD0j4dZA_4",
-//   authDomain: "vaccheck-6a24b.firebaseapp.com",
-//   projectId: "vaccheck-6a24b",
-//   storageBucket: "vaccheck-6a24b.appspot.com",
-//   messagingSenderId: "685620451959",
-//   appId: "1:685620451959:web:8b1946844a5c2dc1550472",
-//   measurementId: "G-BG18CF3S0E"
-// };
-
-// const fire = firebase.initializeApp(firebaseConfig);
-// // const analytics = getAnalytics(app);
-
-// export default fire;
